+++ packages/server/rollup.config.ts
@@ -15,8 +15,6 @@ export const input = [
   'src/rpc/index.ts',
   'src/observable/index.ts',
   'src/subscription.ts',
-  'src/transformer.ts',
-  'src/internals/index.ts',
   'src/shared/index.ts',
 ];
