+++ packages/server/src/http/resolveHTTPResponse.ts
@@ -8,16 +8,12 @@ import {
 } from '../core';
 import { TRPCError, getTRPCErrorFromUnknown } from '../error/TRPCError';
 import { getCauseFromUnknown } from '../error/utils';
-import { transformTRPCResponse } from '../internals/transformTRPCResponse';
 import { TRPCResponse } from '../rpc';
+import { transformTRPCResponse } from '../shared/transformTRPCResponse';
 import { Maybe } from '../types';
 import { getHTTPStatusCode } from './getHTTPStatusCode';
-import {
-  HTTPBaseHandlerOptions,
-  HTTPHeaders,
-  HTTPRequest,
-  HTTPResponse,
-} from './internals/types';
+import { HTTPHeaders, HTTPResponse } from './internals/types';
+import { HTTPBaseHandlerOptions, HTTPRequest } from './types';
 
 const HTTP_METHOD_PROCEDURE_TYPE_MAP: Record<
   string,
