--- packages/server/src/internals/index.ts
@@ -1,3 +1,0 @@
-export type { BaseHandlerOptions, OnErrorFunction } from './types';
-
-export { transformTRPCResponse } from './transformTRPCResponse';
