+++ packages/server/src/shared/index.ts
@@ -1,5 +1,6 @@
 export * from './createProxy';
 export * from './jsonify';
+export * from './transformTRPCResponse';
 
 // For `.d.ts` files https://github.com/trpc/trpc/issues/3943
 export type { SerializeObject } from './internal/serialize';
